+++ CHANGELOG.rst
@@ -9,6 +9,11 @@ Version 3.7
 * GDAL raster metadata can be modified (resolves #181)
 * map canvas preserves scale on window resize (#409)
 * Reclassify Tool: can save and reload the class mapping, fixed (#501)
+* several improvements to SpectralLibrary, e.g. to edit SpectralProfile values
+* adds 'format_py' to QgsExpressionWidget to format string like in python
+* fixes some bugs in imageCube app
+
+
 
 Version 3.6
 -----------